src/workshop-registration/requirements.txt
PyYAML
python-dotenv # having trouble installing in requirements.txt
pandas
requests
streamlit
types-requests
types-PyYAML